Logout, basic database implementation
Logout button added, basic database implementation of Firestore with a template for users and figurines now. Sign up authentication now done locally as it made more sense, Firebase authentication left for login screen. Added usernames to logins and all logins are now saved on the Database to allow the use for usernames. Attempted implementation of profile pictures being saved to the database but did not get it to work. Took out attempted new logic and app runs as usual again.