Library Database Implementation (#6)
Library Database Implementation. Firebase no longer offer Firebase storage for free so had to create workarounds. Solution found was to create a storage on Cloudinary and have it create URL's to save onto Firestore and then fetched from the app. Was originally struggling to make the images appear only to find out fetching the URL creates whitespace, and and once trimmed, appears as it should. More comments added to the existing code for self clarity