Full ML Implementation
Image scanning of PopMarts now completed but the model only contains data for 4 PopMarts. The model was previously converted into a quantised form in order to make it as lightweight as possible, however the model only showed confidence levels for 1 figure. As multiple figures must be scanned, I converted the model into floating point which fixed all issues but required the recoding of the interpreter and model conversion settings