implementation of collection-view (#11)
implementation of collection-view in the display of the figurines. The figures now display more than 1 PopMart at a time and it can be scrolled horizontally too see. Required the revamping of SkullPandaController and LibraryScreen as well as the addition of the cell files. General clean up was done in some modules, mainly in AppDelegate and ProfileScreen due to warnings showing up.